style tweaks

+++ src/app/components/recipe/recipe-view/recipe-view.component.ts
@@ -21,8 +21,11 @@ export class RecipeViewComponent implements OnInit {
   }
 
   delete() {
-    this.recipeService.deleteRecipe(this.recipe._id).then(() => {
-      this.router.navigate(['recipes']);
-    });
+    const confirmed = confirm('Are you sure you want to delete this recipe? This cannot be undone.');
+    if (confirmed) {
+      this.recipeService.deleteRecipe(this.recipe._id).then(() => {
+        this.router.navigate(['recipes']);
+      });
+    }
   }
 }
